+++ src/main/java/memoranda/AuthenticationService/LoginReturns.java
@@ -1,5 +1,9 @@
 package memoranda.AuthenticationService;
 
+/**
+ * @Author Ryan Dinaro
+ * All possible states of login requests. String representation is a short description of the state
+ */
 public enum LoginReturns {
     LOGIN_SUCCESSFUL("User has successfully logged in"),
     USER_NOT_FOUND("User information not found, try again or create a new account"),
@@ -7,12 +11,20 @@ public enum LoginReturns {
     CREATED_ACCOUNT_INFO("User account has been successfully created"),
     USERNAME_TAKEN("This username has already been taken"),
     PASSWORD_CHANGED("The password has successfully been changed");
-    private String statement;
-    LoginReturns(String returnStatement) {
-        statement = returnStatement;
+    private final String stateOfLoginRequest;
+
+    /**
+     * @param state - Description of state
+     */
+    LoginReturns(String state) {
+        stateOfLoginRequest = state;
     }
+
+    /**
+     * @return - the description of the state
+     */
     @Override
     public String toString(){
-        return statement;
+        return stateOfLoginRequest;
     }
 }
